feat(dashboard): implement admin access level

+++ src/server/actions/dashboard/optionListManager/deleteOptionList.ts
@@ -2,12 +2,14 @@
 
 import { prisma } from "@/services/prisma";
 import { revalidatePath } from "next/cache";
+import requireAdminSession from "@/server/services/helpers/auth/requireAdminSession";
 
 export type DeleteOptionListInput = {
   id: number;
 };
 
 const deleteOptionList = async ({ id }: DeleteOptionListInput) => {
+  await requireAdminSession();
   try {
     await prisma.optionList.delete({
       where: {
